rtmp-services: Add "More Info" fields to HLS service
Add fields corresponding to the "More Info" button, such as
"show_more_info" boolean and "more_info_url" to specify parameters
of the "More Info" button needed for its implementation in
window-basic-auto-config.cpp. Use a dummy link for more_info_url for
the moment until we have the actual one.