rtmp-services: Add more_info_url to obs_properties
The function rtmp_common_properties adds a variety
of properties to the obs_properties_t object used
in many of the UI files. Add in a more_info_url
text field to these properties. When a service is
selected, fill out the server list for each service
as well as the more_info_url, so that it can later
be accessed from the settings.